fixed socketry: now writing 1 byte in procedure list instead of 4 bytes

+++ src/Socketry/socketry/Socketry.cs
@@ -23,7 +23,7 @@ namespace socketry
             {
                 Console.WriteLine(procedureName);
                 binaryWriter.Write(Encoding.UTF8.GetBytes(procedureName));
-                binaryWriter.Write(0);
+                binaryWriter.Write((byte)0);
             }
             buffer.Position = 0;
             byte[] result = buffer.ToArray();
